cleaned circular reference
