+++ IPython/testing/messagespec_common.py
@@ -12,7 +12,6 @@ message specification.
 import re
 import sys
 from distutils.version import LooseVersion as V
-from importlib import import_module
 
 import nose.tools as nt
 
@@ -72,22 +71,3 @@ class MimeBundle(Reference):
         for k,v in iteritems(new):
             assert mime_pat.match(k)
             nt.assert_is_instance(v, string_types)
-
-# method to get the correct message spec version for the test
-def get_message_spec_validator(spec_version):
-    
-    """
-    
-    For now there is only one version of the message specification that can be tested.
-    As more versions are added this will be expanded to import the correct version to 
-    be used for the kernel tests.
-    
-    The choice will be made based on spec_version
-    
-    Every version of message spec should implement a validation method that takes 3
-    parameters message, message type and a parent. The message type and parent are
-    optional for validation.
-    
-    """
-    spec_module = import_module('IPython.testing.messagespec')
-    return spec_module.validate_message
